requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.49.0
requests==2.32.5
pydantic==2.11.7
pydantic-settings==2.10.1
openai==1.99.9
qdrant-client==1.15.1
pypdf==5.9.0
python-docx==1.2.0
python-multipart==0.0.20
pytest==8.4.1
